Updated BlenderApplication to act as a proper abstract class
OS specific class instance will get loaded on creation of instanced application.
