+++ blender_applications/darwin_blender_application.py
@@ -24,6 +24,9 @@ class DarwinBlenderApplication(BlenderApplication):
     """
     Darwin (MACOS) Implementation of BlenderApplication
     """
+    def __init__(self, argv):
+        super().__init__(argv)
+
     def __on_focus_object_changed(self, focus_object: QObject):
         """
