+++ blender_applications/win32_blender_application.py
@@ -26,6 +26,9 @@ class Win32BlenderApplication(BlenderApplication):
     """
     Windows implementation of BlenderApplication
     """
+    def __init__(self, argv):
+        super().__init__(argv)
+
     def __on_focus_object_changed(self, focus_object: QObject):
         """
